fix: fix export
